Update Allocation Logics and Resolve Minor Issues (#2)
* add: test code for UntitledHub

* feat: Bank's cancel operations, reallocate, updateallocate

* add: test code for Bank

* fix: remove allocation params of addMarket

* feat: consider available market liquidity when withdraw

* feat: allocations functions of CoreBank

* refactor: code size optimization

+++ src/core/CoreBankFactory.sol
@@ -31,7 +31,7 @@ contract CoreBankFactory {
         IERC20 asset,
         string memory name,
         string memory symbol,
-        uint256 minDelay,
+        uint32 minDelay,
         address initialAdmin
     ) external onlyOwner returns (CoreBank) {
         CoreBank newCoreBank = new CoreBank(
